implement test to find UserEntity by id

+++ src/main/java/com/softserveinc/dokazovi/repositories/UserRepository.java
@@ -18,7 +18,7 @@ import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<UserEntity, Integer> {
 
-	Optional<UserEntity> findById (Long userId);
+	Optional<UserEntity> findById (Integer userId);
 
 	/**
 	 * Gets the user by its email.
